[FE-166] DRY fetch data queries.
